+++ mayan/apps/file_caching/apps.py
@@ -47,7 +47,7 @@ class FileCachingConfig(MayanAppConfig):
         )
 
         SourceColumn(
-            attribute='label', is_identifier=True, source=Cache
+            attribute='label', is_identifier=True, is_object_absolute_url=True, source=Cache
         )
         SourceColumn(
             attribute='get_maximum_size_display', include_label=True,