+++ mayan/apps/file_caching/forms.py
@@ -1,0 +1,9 @@
+from mayan.apps.views.forms import DetailForm
+
+from .models import Cache
+
+
+class CacheDetailForm(DetailForm):
+    class Meta:
+        fields = ()
+        model = Cache