+++ mayan/apps/file_caching/models.py
@@ -6,6 +6,7 @@ from django.core.files.base import ContentFile
 from django.db import models, transaction
 from django.db.models import Sum
 from django.template.defaultfilters import filesizeformat
+from django.urls import reverse
 from django.utils.encoding import force_text
 from django.utils.functional import cached_property
 from django.utils.text import format_lazy
@@ -41,6 +42,13 @@ class Cache(models.Model):
 
     def __str__(self):
         return force_text(s=self.label)
+
+    def get_absolute_url(self):
+        return reverse(
+            viewname='file_caching:cache_detail', kwargs={
+                'cache_id': self.pk
+            }
+        )
 
     def get_files(self):
         return CachePartitionFile.objects.filter(partition__cache__id=self.pk)